pyproject.toml
[tool.poetry]
name = "nutriai"
version = "0.1.0"
description = ""
authors = ["Bruno Gonzaga Santos <brunog86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^0.2.3"
langchain-community = "^0.3.3"
langchainhub = "^0.1.21"
langgraph = "^0.2.39"
langchain-core = "^0.3.12"
langchain = "^0.3.4"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
